requirements.txt
bidict==0.22.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-SocketIO==5.3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
Werkzeug==2.3.6
